pyproject.toml
[tool.poetry]
name = "pjecz_rrhh_personal_api_key"
version = "2.0.0"
description = "API con autentificación para realizar operaciones con la base de datos de RRHH Personal. Hecho con FastAPI."
authors = ["Guillermo Valdes <guillermo@movimientolibre.com>"]
license = "AGPL3"

[tool.poetry.dependencies]
python = "^3.11"
cryptography = "^41.0.2"
fastapi = "^0.100.0"
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.5"}
google-auth = "^2.22.0"
google-cloud = "^0.34.0"
google-cloud-secret-manager = "^2.16.2"
google-cloud-storage = "^2.10.0"
gunicorn = "^20.1.0"
hashids = "^1.3.1"
psycopg2-binary = "^2.9.6"
pydantic = "^2.0.2"
pydantic-settings = "^2.0.1"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
sqlalchemy = "^2.0.18"
sqlalchemy-utils = "^0.41.1"
unidecode = "^1.3.6"
uvicorn = "^0.22.0"

[tool.poetry.dev-dependencies]
black = "^23.7.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
pylint = "^2.17.4"
pylint-sqlalchemy = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 256

[tool.isort]
line_length = 256
profile = "black"
